backend/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.115.4
uvicorn[standard]==0.34.0
starlette==0.41.0

# Pydantic for data validation
pydantic==2.11.1

# Authentication and JWT
pyjwt==2.9.0
passlib==1.7.4

# MongoDB integration
motor==3.6.0
pymongo==4.9.1

# Transformers and AI/ML dependencies (CPU version of FAISS)
transformers==4.50.3
torch==2.6.0
torchvision==0.21.0
sentence-transformers==4.0.1
flagembedding==1.3.4
faiss-cpu==1.8.0

# LangChain
langchain==0.3.22
langchain-community==0.3.7

# HTTP requests and utils
requests==2.32.3
httpx==0.28.1

# Google OAuth2
google-auth==2.35.0
google-auth-oauthlib
google-auth-httplib2

# Env + Crypto
python-dotenv==1.0.1
bcrypt==4.0.1

# File handling
python-multipart==0.0.12
pillow==11.1.0

# Misc
regex==2024.11.6
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
matplotlib==3.9.0
plotly==5.22.0
optimum==1.17.0
auto-gptq==0.6.2
langchain-huggingface==0.0.4

langchain-cli
